+++ MapboxAndroidDemo/src/main/AndroidManifest.xml
@@ -411,8 +411,8 @@
                 android:value="com.mapbox.mapboxandroiddemo.MainActivity" />
         </activity>
         <activity
-            android:name=".examples.styles.GeoJsonClusteringActivity"
-            android:label="@string/activity_styles_data_clusters_title">
+            android:name=".examples.dds.GeoJsonClusteringActivity"
+            android:label="@string/activity_styles_dds_geojson_clusters_title">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value="com.mapbox.mapboxandroiddemo.MainActivity" />
