+++ MapboxAndroidDemo/src/main/res/layout/activity_geojson_clustering.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.mapbox.mapboxandroiddemo.examples.styles.GeoJsonClusteringActivity">
+    tools:context="com.mapbox.mapboxandroiddemo.examples.dds.GeoJsonClusteringActivity">
 
 
     <com.mapbox.mapboxsdk.maps.MapView
