+++ MapboxAndroidDemo/src/main/res/values/descriptions_strings.xml
@@ -19,7 +19,7 @@
     <string name="activity_styles_add_wms_source_description">Adding an external Web Map Service layer to the map.</string>
     <string name="activity_styles_zoom_dependent_fill_color_description">Make a property depend on the map zoom level, in this case, the water layers fill color.</string>
     <string name="activity_dds_create_hotspots_points_description">Use the Mapbox Maps SDK to visualize point data as hotspots.</string>
-    <string name="activity_styles_create_data_cluster_description">Use GeoJSON to visualize point data in clusters.</string>
+    <string name="activity_styles_dds_geojson_clusters_description">Use GeoJSON to visualize point data in clusters.</string>
     <string name="activity_styles_line_layer_description">Create a GeoJSON line source, style it using properties, and add the layer to the map.</string>
     <string name="activity_styles_basic_symbol_layer_description">Display markers on the map by adding a symbol layer. Query the map and animate the icon size if clicked on.</string>
     <string name="activity_styles_local_style_or_raster_source_description">Load a locally stored map style JSON file or custom raster style via a URL.</string>
