+++ MapboxAndroidDemo/src/main/res/values/titles_strings.xml
@@ -18,7 +18,7 @@
     <string name="activity_styles_add_wms_source_title">Add a WMS source</string>
     <string name="activity_styles_zoom_dependent_fill_color_title">Color dependent on zoom level</string>
     <string name="activity_dds_create_hotspots_points_title">Create hotspots from points</string>
-    <string name="activity_styles_data_clusters_title">Create and style data clusters</string>
+    <string name="activity_styles_dds_geojson_clusters_title">Create and style data clusters</string>
     <string name="activity_styles_line_layer_title">Create a line layer</string>
     <string name="activity_styles_basic_symbol_layer_title">Marker symbol layer</string>
     <string name="activity_styles_local_style_or_raster_source_title">Local style or custom raster style</string>
